Restructured parametes and enums are turned into enum classes

+++ src/Views/CustomMenuBar.cpp
@@ -31,37 +31,37 @@ PopupMenu CustomMenuBar::getMenuForIndex(int topLevelMenuIndex, const String &me
 
 	if (topLevelMenuIndex == 0) // Woodwinds
 	{
-		menu.addItem(Woodwinds::Piccolo, PiccoloName);
-		menu.addItem(Woodwinds::Flute, FluteName);
-		menu.addItem(Woodwinds::Oboe, OboeName);
-		menu.addItem(Woodwinds::CorAnglais, CorAnglaisName);
-		menu.addItem(Woodwinds::Clarinet, ClarinetName);
-		menu.addItem(Woodwinds::BassClarinet, BassClarinetName);
-		menu.addItem(Woodwinds::Bassoon, BassoonName);
-		menu.addItem(Woodwinds::Contrabassoon, ContrabassoonName);
+		menu.addItem(static_cast<int>(Woodwinds::Piccolo), PiccoloName);
+		menu.addItem(static_cast<int>(Woodwinds::Flute), FluteName);
+		menu.addItem(static_cast<int>(Woodwinds::Oboe), OboeName);
+		menu.addItem(static_cast<int>(Woodwinds::CorAnglais), CorAnglaisName);
+		menu.addItem(static_cast<int>(Woodwinds::Clarinet), ClarinetName);
+		menu.addItem(static_cast<int>(Woodwinds::BassClarinet), BassClarinetName);
+		menu.addItem(static_cast<int>(Woodwinds::Bassoon), BassoonName);
+		menu.addItem(static_cast<int>(Woodwinds::Contrabassoon), ContrabassoonName);
 	}
 	else if (topLevelMenuIndex == 1) // Brass
 	{
-		menu.addItem(Brass::FrenchHorn, FrenchHornName);
-		menu.addItem(Brass::Trumpet, TrumpetName);
-		menu.addItem(Brass::TenorTrombone, TenorTromboneName);
-		menu.addItem(Brass::BassTrombone, BassTromboneName);
-		menu.addItem(Brass::Cimbasso, CimbassoName);
-		menu.addItem(Brass::Tuba, TubaName);
+		menu.addItem(static_cast<int>(Brass::FrenchHorn), FrenchHornName);
+		menu.addItem(static_cast<int>(Brass::Trumpet), TrumpetName);
+		menu.addItem(static_cast<int>(Brass::TenorTrombone), TenorTromboneName);
+		menu.addItem(static_cast<int>(Brass::BassTrombone), BassTromboneName);
+		menu.addItem(static_cast<int>(Brass::Cimbasso), CimbassoName);
+		menu.addItem(static_cast<int>(Brass::Tuba), TubaName);
 	}
 	else if (topLevelMenuIndex == 2) // Strings
 	{
-		menu.addItem(Strings::Violin, ViolinName);
-		menu.addItem(Strings::Viola, ViolaName);
-		menu.addItem(Strings::Violoncello, VioloncelloName);
-		menu.addItem(Strings::DoubleBass, DoubleBassName);
+		menu.addItem(static_cast<int>(Strings::Violin), ViolinName);
+		menu.addItem(static_cast<int>(Strings::Viola), ViolaName);
+		menu.addItem(static_cast<int>(Strings::Violoncello), VioloncelloName);
+		menu.addItem(static_cast<int>(Strings::DoubleBass), DoubleBassName);
 	}
 	else if (topLevelMenuIndex == 3) // Percussion
 	{
-		menu.addItem(Percussion::Celeste, CelesteName);
-		menu.addItem(Percussion::Harp, HarpName);
-		menu.addItem(Percussion::Timpani, TimpaniName);
-		menu.addItem(Percussion::Marimba, MarimbaName);
+		menu.addItem(static_cast<int>(Percussion::Celeste), CelesteName);
+		menu.addItem(static_cast<int>(Percussion::Harp), HarpName);
+		menu.addItem(static_cast<int>(Percussion::Timpani), TimpaniName);
+		menu.addItem(static_cast<int>(Percussion::Marimba), MarimbaName);
 	}
 
 	return menu;
